requirements.txt
allure_python_commons==2.13.5
pytest==8.2.0
selenium==4.23.1
